Headers cleanup.
