+++ pcsx2/PAD/Linux/KeyStatus.cpp
@@ -14,6 +14,7 @@
  */
 
 #include "KeyStatus.h"
+#include "controller.h"
 
 void KeyStatus::Init()
 {
