+++ pcsx2/PAD/Linux/PAD.h
@@ -15,41 +15,8 @@
 
 #pragma once
 
-#define GAMEPAD_NUMBER 2 // numbers of gamepad
-
-#include <wx/string.h>
-#include <wx/tokenzr.h>
-#include <wx/intl.h>
-#include <wx/log.h>
-#include <wx/filename.h>
-#include "Utilities/pxStreams.h"
-#include "Utilities/Console.h"
-#include "App.h"
-#include <stdio.h>
-#include <assert.h>
-
-#ifdef _WIN32
-#include <windows.h>
-#include <windowsx.h>
-
-#else
-/*
-#include <X11/Xlib.h>
-#include <X11/Xutil.h>
-#include <X11/keysym.h>
-*/
-
-#endif
-
-#include <array>
-#include <vector>
-#include <map>
-#include <string>
-#include <memory>
-#include <mutex>
-#include <queue>
-
-#define PADdefs
+#include "Global.h"
+#include "mt_queue.h"
 
 enum PadOptions
 {
@@ -62,62 +29,8 @@ enum PadOptions
 	PADOPTION_MOUSE_R = 0x40,
 };
 
-enum PadCommands
-{
-	CMD_SET_VREF_PARAM = 0x40,
-	CMD_QUERY_DS2_ANALOG_MODE = 0x41,
-	CMD_READ_DATA_AND_VIBRATE = 0x42,
-	CMD_CONFIG_MODE = 0x43,
-	CMD_SET_MODE_AND_LOCK = 0x44,
-	CMD_QUERY_MODEL_AND_MODE = 0x45,
-	CMD_QUERY_ACT = 0x46,  // ??
-	CMD_QUERY_COMB = 0x47, // ??
-	CMD_QUERY_MODE = 0x4C, // QUERY_MODE ??
-	CMD_VIBRATION_TOGGLE = 0x4D,
-	CMD_SET_DS2_NATIVE_MODE = 0x4F // SET_DS2_NATIVE_MODE
-};
-
-enum gamePadValues
-{
-	PAD_L2 = 0,   // L2 button
-	PAD_R2,       // R2 button
-	PAD_L1,       // L1 button
-	PAD_R1,       // R1 button
-	PAD_TRIANGLE, // Triangle button ▲
-	PAD_CIRCLE,   // Circle button ●
-	PAD_CROSS,    // Cross button ✖
-	PAD_SQUARE,   // Square button ■
-	PAD_SELECT,   // Select button
-	PAD_L3,       // Left joystick button (L3)
-	PAD_R3,       // Right joystick button (R3)
-	PAD_START,    // Start button
-	PAD_UP,       // Directional pad ↑
-	PAD_RIGHT,    // Directional pad →
-	PAD_DOWN,     // Directional pad ↓
-	PAD_LEFT,     // Directional pad ←
-	PAD_L_UP,     // Left joystick (Up) ↑
-	PAD_L_RIGHT,  // Left joystick (Right) →
-	PAD_L_DOWN,   // Left joystick (Down) ↓
-	PAD_L_LEFT,   // Left joystick (Left) ←
-	PAD_R_UP,     // Right joystick (Up) ↑
-	PAD_R_RIGHT,  // Right joystick (Right) →
-	PAD_R_DOWN,   // Right joystick (Down) ↓
-	PAD_R_LEFT    // Right joystick (Left) ←
-};
-
-#if defined(__unix__) || defined(__APPLE__)
-#include "GamePad.h"
-#endif
-#include "bitwise.h"
-#include "controller.h"
-#include "KeyStatus.h"
-#include "mt_queue.h"
-
 extern FILE* padLog;
 extern void initLogging();
-
-//#define PAD_LOG __Log
-//#define PAD_LOG __LogToConsole
 
 extern keyEvent event;
 extern MtQueue<keyEvent> g_ev_fifo;
@@ -125,10 +38,6 @@ extern MtQueue<keyEvent> g_ev_fifo;
 s32 _PADopen(void* pDsp);
 void _PADclose();
 void PADsetMode(int pad, int mode);
-
-void __LogToConsole(const char* fmt, ...);
-void PADLoadConfig();
-void PADSaveConfig();
 
 void SysMessage(char* fmt, ...);
