+++ pcsx2/PAD/Linux/SDL/joystick.h
@@ -18,9 +18,9 @@
 #include <SDL.h>
 #include <SDL_haptic.h>
 
-#include "PAD/Linux/GamePad.h"
-#include "PAD/Linux/PAD.h"
-#include "PAD/Linux/controller.h"
+#include "../Global.h"
+#include "../GamePad.h"
+#include "../controller.h"
 #define NB_EFFECT 2 // Don't use more than two, ps2 only has one for big motor and one for small(like most systems)
 // holds all joystick info
 class JoystickInfo : public GamePad
