+++ pcsx2/PAD/Linux/wx_dialog/GamepadConfiguration.h
@@ -20,9 +20,9 @@
 
 #include <wx/wx.h>
 
-#include "PAD/Linux/GamePad.h"
-#include "PAD/Linux/keyboard.h"
-#include "PAD/Linux/PAD.h"
+#include "../GamePad.h"
+#include "../keyboard.h"
+#include "../Global.h"
 
 static const s32 rumble_slider_id = wxID_HIGHEST + 200 + 1;
 static const s32 joy_slider_id = wxID_HIGHEST + 200 + 2;
